requirements.txt
python-telegram-bot==20.7
python-dotenv
pydrive
gspread
google-auth
python-dotenv
